+++ lib/vertical_timeline_widget.dart
@@ -8,3 +8,4 @@ library;
 export 'src/timeline.dart';
 export 'src/timeline_data.dart';
 export 'src/timeline_theme.dart';
+export 'src/steps.dart';